refactor score submission endpoint
