+++ app/usecases/beatmaps.py
@@ -172,4 +172,18 @@ async def update_status(beatmap: Beatmap, new_status: RankedStatus) -> None:
     await repositories.beatmaps.update_status(beatmap.id, new_status)
 
 
+async def update_playcounts(beatmap: Beatmap, increment_passes: bool) -> None:
+    """Update a beatmaps playcounts in cache and the database."""
+    beatmap.plays += 1
+
+    if increment_passes:
+        beatmap.passes += 1
+
+    await repositories.beatmaps.update_playcounts(
+        beatmap.id,
+        beatmap.plays,
+        beatmap.passes,
+    )
+
+
 # delete
